Introduce error modules to locktime
In order to de-clutter the public API introduce public `error` modules
to both locktime modules and move all error types there (in `units`
crate).

As is now policy, re-export error types that appear in directly the
public API but don't doc inline them. This is a bit confusing because
_all_ the error types except `ParseError` appear directly in the API.
